+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- Add a new `slideOnWheel` option to slide to the next or previous image by wheeling the mouse, with optional modifier key support (e.g. `ctrl`) (#528).
+- Enhance the `zoomOnWheel` option to support requiring a modifier key (e.g. `ctrl`) while wheeling to zoom (#528).
+
 ## 1.13.0 (Sep 6, 2026)
 
 - Support rotating images with two-finger touch gestures using the `rotateOnTouch` option.